pyproject.toml
[project]
name = "activity-logs-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiokafka>=0.12.0",
    "argon2-cffi>=25.1.0",
    "fastapi[all]>=0.124.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
]
